some refactoring and theme changes
