+++ lua/custom/plugins/neoscroll.lua
@@ -1,0 +1,7 @@
+return {
+  "karb94/neoscroll.nvim",
+  config = function ()
+    require('neoscroll').setup {}
+  end
+}
+
