+++ lua/custom/plugins/rose-pine.lua
@@ -1,0 +1,103 @@
+return {
+  'rose-pine/neovim',
+  name = 'rose-pine',
+
+  config = function()
+    require('rose-pine').setup {
+      variant = 'moon', -- auto, main, moon, or dawn
+      dark_variant = 'moon', -- main, moon, or dawn
+      dim_inactive_windows = false,
+      extend_background_behind_borders = true,
+
+      styles = {
+        bold = true,
+        italic = true,
+        transparency = false,
+      },
+
+      groups = {
+        border = 'muted',
+        link = 'iris',
+        panel = 'surface',
+
+        error = 'love',
+        hint = 'iris',
+        info = 'foam',
+        warn = 'gold',
+
+        git_add = 'foam',
+        git_change = 'rose',
+        git_delete = 'love',
+        git_dirty = 'rose',
+        git_ignore = 'muted',
+        git_merge = 'iris',
+        git_rename = 'pine',
+        git_stage = 'iris',
+        git_text = 'rose',
+        git_untracked = 'subtle',
+
+        headings = {
+          h1 = 'iris',
+          h2 = 'foam',
+          h3 = 'rose',
+          h4 = 'gold',
+          h5 = 'pine',
+          h6 = 'foam',
+        },
+        -- Alternatively, set all headings at once.
+        -- headings = "subtle",
+      },
+
+      highlight_groups = {
+        -- Comment = { fg = "foam" },
+        -- VertSplit = { fg = "muted", bg = "muted" },
+      },
+
+      before_highlight = function(group, highlight, palette)
+        -- Disable all undercurls
+        -- if highlight.undercurl then
+        --     highlight.undercurl = false
+        -- end
+        --
+        -- Change palette colour
+        -- if highlight.fg == palette.pine then
+        --     highlight.fg = palette.foam
+        -- end
+      end,
+    }
+
+    vim.cmd 'colorscheme rose-pine'
+  end,
+}
+
+
+  -- {
+  --   'ajmwagar/vim-deus'
+  -- },
+
+  -- {
+  --   -- Theme inspired by Atom
+  --   'navarasu/onedark.nvim',
+  --   priority = 1000,
+  --   config = function()
+  --     vim.cmd.colorscheme 'onedark'
+  --   end,
+  -- },
+  --
+  -- {
+  --   -- Set lualine as statusline
+  --   'nvim-lualine/lualine.nvim',
+  --   -- See `:help lualine.txt`
+  --   opts = {
+  --     options = {
+  --       icons_enabled = false,
+  --       theme = 'onedark',
+  --       component_separators = '|',
+  --       section_separators = '',
+  --     },
+  --   },
+  -- },
+
+
+
+
